http-api/requirements.txt
Flask~=2.3.2
pyyaml~=6.0.1
flask_restful~=0.3.10
